feat: restore bounding box algorihm for cell location and cleanup old code
See comment at https://codeberg.org/beacondb/beacondb/pulls/134#issuecomment-6193675
In short, the weighted average algorithm is not adapted to locate users
using cell towers. A lot of problems come into account, with the most
important problems being the non-uniform coverage leading to incorrect
locations, and users being always out of the accuracy range, as the
accuracy is the accuracy of the cell tower, not its coverage area.

For these reasons, I prefer using bounding box for cell location until
a better algorithm is implemented for this specific case.