#8 Report parse errors.
Removes the Option<> for the data on the geolocate service which causes Actix to
trigger a 400 Bad Request when the JSON cannot be parsed. The error reported
back to the client will be the actual serde error, e.g:
`Json deserialize error: missing field `cellId` at line 1 column 231`

Further customization of the errors could be possible, but it looks like the
errors from Serde are quite useful already.